refactor/#13: 사이즈 종류 선택버튼 컴포넌트화
